refactor: extract QuestionAnswerOptions and QuestionCard
