+++ packages/smooth_app/lib/pages/hunger_games/question_page.dart
@@ -4,40 +4,27 @@ import 'package:matomo_tracker/matomo_tracker.dart';
 import 'package:openfoodfacts/openfoodfacts.dart';
 import 'package:openfoodfacts/utils/OpenFoodAPIConfiguration.dart';
 import 'package:provider/provider.dart';
-import 'package:shimmer/shimmer.dart';
-import 'package:smooth_app/cards/product_cards/product_image_carousel.dart';
-import 'package:smooth_app/cards/product_cards/product_title_card.dart';
 import 'package:smooth_app/database/local_database.dart';
 import 'package:smooth_app/generic_lib/design_constants.dart';
 import 'package:smooth_app/generic_lib/duration_constants.dart';
 import 'package:smooth_app/generic_lib/loading_dialog.dart';
 import 'package:smooth_app/helpers/robotoff_insight_helper.dart';
 import 'package:smooth_app/pages/hunger_games/congrats.dart';
+import 'package:smooth_app/pages/hunger_games/question_answers_options.dart';
+import 'package:smooth_app/pages/hunger_games/question_card.dart';
 import 'package:smooth_app/query/product_questions_query.dart';
 import 'package:smooth_app/query/questions_query.dart';
 import 'package:smooth_app/widgets/smooth_scaffold.dart';
-
-const Color _noBackground = Colors.redAccent;
-const Color _yesBackground = Colors.lightGreen;
-const Color _yesNoTextColor = Colors.white;
-
-const List<InsightType> ALL_INSIGHTS = <InsightType>[
-  InsightType.CATEGORY,
-  InsightType.LABEL,
-  InsightType.PRODUCT_WEIGHT,
-  InsightType.PACKAGER_CODE,
-  InsightType.BRAND,
-];
 
 class QuestionPage extends StatefulWidget {
   const QuestionPage({
     this.product,
     this.questions,
     this.updateProductUponAnswers,
-    this.insightTypes = ALL_INSIGHTS,
+    this.insightTypes,
   });
 
-  final List<InsightType> insightTypes;
+  final List<InsightType>? insightTypes;
   final Product? product;
   final List<RobotoffQuestion>? questions;
   final Function()? updateProductUponAnswers;
@@ -181,230 +168,20 @@ class _QuestionPageState extends State<QuestionPage>
     }
 
     final RobotoffQuestion question = questions[questionIndex];
+    final AppLocalizations appLocalizations = AppLocalizations.of(context);
 
     return Column(
       children: <Widget>[
-        _buildQuestionCard(
-          context,
+        QuestionCard(question),
+        QuestionAnswersOptions(
           question,
-        ),
-        _buildAnswerOptions(
-          context,
-          question,
+          onAnswer: (InsightAnnotation answer) => trySave(
+            question,
+            answer,
+            appLocalizations,
+          ),
         )
       ],
-    );
-  }
-
-  Widget _buildQuestionCard(
-    BuildContext context,
-    RobotoffQuestion question,
-  ) {
-    final Future<Product> productFuture = OpenFoodAPIClient.getProduct(
-      ProductQueryConfiguration(question.barcode!),
-    ).then((ProductResult result) => result.product!);
-
-    final Size screenSize = MediaQuery.of(context).size;
-
-    return FutureBuilder<Product>(
-        future: productFuture,
-        builder: (BuildContext context, AsyncSnapshot<Product> snapshot) {
-          if (!snapshot.hasData) {
-            return _buildQuestionShimmer();
-          }
-          final Product product = snapshot.data!;
-          return Card(
-            elevation: 4,
-            clipBehavior: Clip.antiAlias,
-            shape: const RoundedRectangleBorder(
-              borderRadius: ROUNDED_BORDER_RADIUS,
-            ),
-            child: Column(
-              children: <Widget>[
-                ProductImageCarousel(
-                  product,
-                  height: screenSize.height / 6,
-                  onUpload: (_) {},
-                ),
-                Padding(
-                  padding: const EdgeInsets.symmetric(horizontal: SMALL_SPACE),
-                  child: Column(
-                    children: <Widget>[
-                      ProductTitleCard(
-                        product,
-                        true,
-                        dense: true,
-                      ),
-                    ],
-                  ),
-                ),
-                _buildQuestionText(context, question),
-              ],
-            ),
-          );
-        });
-  }
-
-  Widget _buildQuestionText(BuildContext context, RobotoffQuestion question) {
-    return Container(
-      color: const Color(0xFFFFEFB7),
-      padding: const EdgeInsets.all(SMALL_SPACE),
-      child: Column(
-        children: <Widget>[
-          Container(
-            alignment: Alignment.center,
-            padding: const EdgeInsetsDirectional.only(bottom: SMALL_SPACE),
-            child: Text(
-              question.question!,
-              style: Theme.of(context)
-                  .textTheme
-                  .headline4!
-                  .apply(color: Colors.black),
-            ),
-          ),
-          Container(
-            decoration: const BoxDecoration(
-              borderRadius: BorderRadius.all(ANGULAR_RADIUS),
-              color: Colors.black,
-            ),
-            padding: const EdgeInsets.all(SMALL_SPACE),
-            child: Text(
-              question.value!,
-              style: Theme.of(context)
-                  .textTheme
-                  .headline4!
-                  .apply(color: Colors.white),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-
-  Widget _buildQuestionShimmer() => Shimmer.fromColors(
-        baseColor: const Color(0xFFFFEFB7),
-        highlightColor: Colors.white,
-        child: Card(
-          elevation: 4,
-          clipBehavior: Clip.antiAlias,
-          shape: const RoundedRectangleBorder(
-            borderRadius: ROUNDED_BORDER_RADIUS,
-          ),
-          child: Container(
-            height: LARGE_SPACE * 10,
-          ),
-        ),
-      );
-
-  Widget _buildAnswerOptions(
-    BuildContext context,
-    RobotoffQuestion question,
-  ) {
-    final double yesNoHeight = MediaQuery.of(context).size.width / (3 * 1.25);
-
-    return Expanded(
-      child: Column(
-        children: <Widget>[
-          Row(
-            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-            children: <Widget>[
-              Expanded(
-                child: SizedBox(
-                  height: yesNoHeight,
-                  child: _buildAnswerButton(
-                    question: question,
-                    insightAnnotation: InsightAnnotation.NO,
-                    backgroundColor: _noBackground,
-                    contentColor: _yesNoTextColor,
-                  ),
-                ),
-              ),
-              Expanded(
-                child: SizedBox(
-                  height: yesNoHeight,
-                  child: _buildAnswerButton(
-                    question: question,
-                    insightAnnotation: InsightAnnotation.YES,
-                    backgroundColor: _yesBackground,
-                    contentColor: _yesNoTextColor,
-                  ),
-                ),
-              ),
-            ],
-          ),
-          const SizedBox(height: SMALL_SPACE),
-          Row(
-            mainAxisAlignment: MainAxisAlignment.center,
-            children: <Widget>[
-              _buildAnswerButton(
-                question: question,
-                insightAnnotation: InsightAnnotation.MAYBE,
-                backgroundColor: const Color(0xFFFFEFB7),
-                contentColor: Colors.black,
-                textButton: true,
-              ),
-            ],
-          ),
-        ],
-      ),
-    );
-  }
-
-  Widget _buildAnswerButton({
-    required RobotoffQuestion question,
-    required InsightAnnotation insightAnnotation,
-    required Color backgroundColor,
-    required Color contentColor,
-    bool textButton = false,
-    EdgeInsets padding = const EdgeInsets.all(VERY_SMALL_SPACE),
-  }) {
-    final AppLocalizations appLocalizations = AppLocalizations.of(context);
-    String buttonText;
-    IconData iconData;
-    switch (insightAnnotation) {
-      case InsightAnnotation.YES:
-        buttonText = appLocalizations.yes;
-        iconData = Icons.check;
-        break;
-      case InsightAnnotation.NO:
-        buttonText = appLocalizations.no;
-        iconData = Icons.clear;
-        break;
-      case InsightAnnotation.MAYBE:
-        buttonText = appLocalizations.skip;
-        iconData = Icons.question_mark;
-    }
-    final Icon buttonIcon = Icon(
-      iconData,
-      color: contentColor,
-      size: 36,
-    );
-    final Text buttonLabel = Text(
-      buttonText,
-      style: Theme.of(context).textTheme.headline2!.apply(color: contentColor),
-    );
-
-    return Padding(
-      padding: padding,
-      child: TextButton.icon(
-        onPressed: () => trySave(
-          question,
-          insightAnnotation,
-          appLocalizations,
-        ),
-        style: textButton
-            ? null
-            : ButtonStyle(
-                backgroundColor: MaterialStateProperty.all(backgroundColor),
-                shape: MaterialStateProperty.all(
-                  const RoundedRectangleBorder(
-                    borderRadius: ROUNDED_BORDER_RADIUS,
-                  ),
-                ),
-              ),
-        icon: buttonIcon,
-        label: buttonLabel,
-      ),
     );
   }
